+++ entry/src/main/config.json
@@ -14,7 +14,7 @@
   },
   "deviceConfig": {},
   "module": {
-    "package": "com.erikagutierrez.multiplemediapicker",
+    "package": "com.deveco.multiplemediapicker",
     "name": ".MyApplication",
     "deviceType": [
       "phone","tv","tablet"
@@ -37,7 +37,7 @@
           }
         ],
         "orientation": "unspecified",
-        "name": "com.erikagutierrez.multiplemediapicker.MainAbility",
+        "name": "com.deveco.multiplemediapicker.MainAbility",
         "icon": "$media:icon",
         "description": "$string:mainability_description",
         "label": "$string:app_name",
