+++ multiplemediapicker/src/main/java/com/deveco/multiplemediapicker/interfaces/FractionClickListner.java
@@ -1,0 +1,15 @@
+package com.deveco.multiplemediapicker.interfaces;
+
+import java.util.List;
+
+/**
+ * FractionClickListner
+ *
+ * @since 2021-08-03
+ */
+
+public interface FractionClickListner {
+
+   void itemClicked(String from, List<String> bitmapList, List<String> bucketNames);
+
+}
