--- multiplemediapicker/src/main/java/com/erikagutierrez/multiplemediapicker/interfaces/FractionClickListner.java
@@ -1,15 +1,0 @@
-package com.erikagutierrez.multiplemediapicker.interfaces;
-
-import java.util.List;
-
-/**
- * FractionClickListner
- *
- * @since 2021-08-03
- */
-
-public interface FractionClickListner {
-
-   void itemClicked(String from, List<String> bitmapList, List<String> bucketNames);
-
-}
