+++ multiplemediapicker/src/test/java/com/deveco/multiplemediapicker/ExampleTest.java
@@ -1,0 +1,21 @@
+package com.deveco.multiplemediapicker;
+
+import org.junit.Test;
+
+/**
+ * ExampleTest .
+ *
+ * @since 2021-08-03
+ */
+
+public class ExampleTest {
+      @Test
+    public void onStart() {
+      // Do nothing .
+    }
+
+   @Override
+   public void onStart() {
+     throw new UnsupportedOperationException();
+    }
+}
