--- multiplemediapicker/src/test/java/com/erikagutierrez/multiplemediapicker/ExampleTest.java
@@ -1,21 +1,0 @@
-package com.erikagutierrez.multiplemediapicker;
-
-import org.junit.Test;
-
-/**
- * ExampleTest .
- *
- * @since 2021-08-03
- */
-
-public class ExampleTest {
-      @Test
-    public void onStart() {
-      // Do nothing .
-    }
-
-   @Override
-   public void onStart() {
-     throw new UnsupportedOperationException();
-    }
-}
